+++ packages/elements/basic-elements/src/index.ts
@@ -3,4 +3,3 @@
  */
 
 export * from './createBasicElementPlugins';
-export * from './types';