+++ packages/elements/basic-elements/src/types.ts
@@ -1,5 +1,0 @@
-import { HeadingPlugin } from '@udecode/plate-heading';
-
-export interface BasicElementsPlugin {
-  heading?: HeadingPlugin;
-}